# Optimize short string parsing (part 2)